Randomize test suite mongod ports and data dirs. (#30)
The test suite’s mongods sometimes clobber each other, which causes
spurious failures.
This changeset makes those mongods use temporary directories and
ephemeral,
kernel-chosen ports in order to minimize conflicts.

Along the way, this upgrades to go 1.19 and adds verification of the
HTTP status codes
in the test.

During review there was concern expressed at the fact that the test
suite initializes
the source as a replSet but doesn’t configure the destination or
metadata cluster
as such. Since this is the existing setup this changeset doesn’t alter
it; REP-2319
documents a follow-up investigation regarding that.