REP-1636: Allow local read concern (#28)
The verifier’s default behavior is to enforce majority read concern on
all connections (source, destination, & metadata). This, however, may
break if a MongoDB cluster lacks support for majority read concern.

This changeset provides a way for users to disable that
majority-enforcement and instead just defer to whatever read concern the
connection uses.