REP-2334: Remove explicit majority-read-concern settings. (#34)
There were 3 places where the verifier set majority read concern on a
individual DB commands. REP-1636 put one of those under a conditional so
that the --ignoreReadConcern flag would cause the verifier to omit it;
however, the two in partitions.go were overlooked.

We could extend the same conditional logic to partitions.go that
REP-1636 added; however, since we set the client-default read concern to
majority anyway, setting the read concern to majority in an individual
command is unnecessary even in the case where we want to enforce
majority read concern.

This changeset, then, removes all 3 cases where we set majority read
concern on an individual DB command, which preserves existing behavior
in the default, force-majority-read-concern behavior while fixing the
behavior in our new, allow-any-read-concern behavior.