REP-2357: Fix change stream race condition (#36)
TestGenerationalRechecking() occasionally failed in Evergreen because
GetVerificationStatus() returned no tasks (at all). The bug was where
the verifier’s change stream thread created document-verification
tasks for the prior generation after GenerateRecheckTasks() already ran
for that generation.

Thus, if a change stream event would cause a document to go from
matching to non-matching, there’s a chance the verifier would miss
that change event and thus report a false match.

This changeset fixes that race condition by making GenerateRecheckTasks
and ClearRecheckDocs happen under the same lock that increments the
generation, and insertRecheckDocs acquire that same lock.

A few additional debugging log statements are added as well, as well as
documentation for a few methods that previously lacked it.