REP-2641 Hide total-data metrics in recheck generations. (#43)
Change events don’t always present a reliable means of discerning
document size. As a result of this, the verifier assumes that every
document it learns about from the change stream is maximally large
(i.e., 16 GiB)

Reasonable as this is for partitioning purposes, it invalidates
REP-1634’s effort at presenting a metric of the total data pending for
a given recheck generation (i.e., generation > 0). Since it’s not
possible to present this metric in a consistently-meaningful way, this
changeset suppresses the metric. It involves a small bit of refactoring
of the display logic (which previously assumed that anywhere a
total-docs metric is available would also present a total-bytes metric)
as well as of the aggregation query that fetches per-namespace
statistics from the verifier’s metadata.