REP-2690 Fix same-generation conflict between change event & failed compare. (#45)
If a failed comparison and a change event happen on the same document
for the same generation, the verifier has been failing because the 2nd
insert into the recheck queue fails due to a duplicate-key error.

This fixes that by making the `replace` command filter only on `_id`,
not `_id` and `DataSize`. A test is added (with some deduplications).