REP-3431 Create partitions using _ids of filtered documents (#7)
This PR adds a filter to the partition logic so that partition bounds are created by randomly sampling filtered documents. Number of partitions is calculated by multiplying a ratio of filtered documents with the entire-collection num of partitions. Since sampling of partition bounds will use filtered documents, they will be more evenly distributed across partitions. A more reasonable number of partitions will also be created if only a small subset of documents out of a collection are to be included in the filter. If the partitioning-after-filtering approach takes too long (timeout error), it will fall back to partitioning without filtering.