REP-5201 Persist the change stream’s resume token. (#30)
This teaches migration-verifier to persist its resume token. Without this, success reports from the verifier aren’t fully reliable because during any downtime it’s possible for change events to happen that would go unverified.

This changeset institutes persistence of the resume token by persisting the token roughly every 10 seconds.

This changeset alters (and simplifies) the verifier’s method for setting its `srcStartAtTs` property. The rationale for this follows:

Before this changeset, the verifier called `GetLastOpTimeAndSyncShardClusterTime()` before opening the change stream. If that function succeeded, the verifier would start the change stream at that time and set it as `srcStartAtTs`; if not, the verifier would ignore the failure, start the change stream, read the resume token’s timestamp as well as the cluster time, then set the earlier of those two as `srcStartAtTs`.

If `min(tokenTs, clusterTs)` is acceptable and always available, then there’s no reason to call `GetLastOpTimeAndSyncShardClusterTime()` (which can fail) in the first place. Thus, this changeset removes that function and always determines `srcStartAtTs` from the new change stream’s resume token (or session cluster time).

If a resume token was previously persisted, then the change stream is initialized with that resume token. Note that `srcStartAtTs` will correspond to that resume token, but that shouldn’t actually affect anything because it’s OK for `srcStartAtTs` to be “in the past” since it’s only use is to be a `find` command’s `readConcern.afterClusterTime`.