REP-5218 Roll back tasks as needed when restarting. (#35)
The worker threads look for status=added tasks when finding work to do. When it grabs a task it sets status=processing; thus, no other threads will try to grab that task.

If the verifier crashes, though, any tasks that were status=processing need to be reset; otherwise no worker thread will ever pick them up, and the verifier will hang.

This changeset implements that reset of status=processing tasks so that the verifier will be able to finish after a restart.

Specifically:
- If the primary task is found `processing`, this means that the initial listing of namespaces was interrupted. Thus, we now reset the primary task to `added`, and all other tasks are deleted.
- If any collection-verification tasks are found `processing`, that means we never finished writing out the collection’s partitions and checking its metadata. Any existing partition tasks are unusuable, so they get deleted, and the collection-verification task is reset to `added`.
- Any remaining `processing` partition tasks are reset to `added`.

The use of transactions to do these rollbacks means that the verifier’s metadata can no longer be a standalone mongod.