REP-5283 Report change event document sizes accurately. (#43)
Previously all change events’ document sizes were recorded “pessimistically” as 16 MiB. This helped to avoid OOMs. It came at a cost, though: when the recheck queue is converted to recheck tasks, those tasks are sized so as to approximate the configured partition size. Thus, if the partition size was 400 MiB (the default), only 25 change events could fit into a recheck task. If there are 250,000 pending rechecks—not unfeasible for a large, busy data set after generation 0—that’s 10,000 tasks to create and perform, which is inefficient.

PR #34 all but eliminates the OOMs, which undercuts that “pessimism”’s benefit. It makes more sense now to allow for the possibility of large recheck tasks in order to minimize the number of tasks. Moreover, we can get pretty good confidence about document sizes from change events anyway:
- Insert & replace events always include the `fullDocument`.
- Update events can be configured to include the current `fullDocument`.
- Delete events refer to a document that probably no longer exists, so we can safely estimate its size to be “small”.

This changeset, then:
1. configures the change stream to include `fullDocument` in update events, and
2. records document sizes from the change event.