Ensure parallelization of rechecks by capping recheck task size. (#45)
Previously the only limit on a recheck task’s size was BSON’s document-size limit (i.e., 16 MiB). Thus, a single task could feasibly include every recheck. If there are millions of rechecks, that means only 1 thread is doing all of them, which increases the likelihood that migration-verifier would fail to keep pace with the source workload.

The present changeset fixes that by limiting the number of documents in a recheck task as per the number of pending rechecks and the number of workers.