REP-5299 Finalize change stream with a timestamp. (#47)
Previously to finalize its change stream the verifier would just TryNext() until no responses arrived. This works fine for unsharded clusters, but for sharded clusters it can potentially miss events.

This changeset adopts more “tried-and-true” logic: grab a timestamp via `appendOplogNote`, then iterate the change stream until the stream’s resume token’s timestamp meets or exceeds that timestamp. As verification of that change, this changeset removes the test skips that REP-5306 (PR #48) imposed for sharded clusters.

The timestamp-grabbing logic is unconditionally retryable, which requires an update to the retryer module to implement. This changeset refactors that module a bit to prevent further duplication.

Additional changes:
- go.sum is now part of the repo.
- Go 1.22 is now a minimum requirement.
- Tests no longer spew output when the worker delay is 0. The log message for that delay is also tweaked to show the delay.
- Mention in code of the defunct “refetch” collection is removed.
- The change stream no longer logs whenever a batch of events arrives. For busy migrations it would be excessive.
- Verifier instances in tests now wait 10ms rather than 15s between generations, which makes a few tests much quicker.
- A new CheckRunner improves the ergonomics of running CheckDriver in tests.
- TestGenerationalRechecking used to verify that an insert after WritesOff() gets handled as a recheck. This doesn’t make a lot of sense since WritesOff() means, by definition, that no more events should happen on the server. Thus, this changeset removes that test, which was proving rather race-prone with the rest of these changes.
- The tests now use majority read & write concern.
- CI now parallelizes a bit more of its setup.
- A new CheckRunner struct smooths over running Check from e2e tests.
- Tests now use a 10ms verificationStatusCheckInterval rather than 15-minute.