REP-5274 Compare full index specs rather than just a subset. (#50)
Previously migration-verifier only checked the parts of index specs that go in a mongo.IndexSpecification. That struct is incomplete, though; it doesn’t, for example, include partial index filters. Thus, migration-verifier didn’t compare those parts of an index specification.

This changeset alters the index-checking logic to check full index specifications. Note that this comparison is rather “coarse” and may flag some specifications as mismatching when they actually produce the same index. For example, `{foo: 1}` and `{foo: 2}` actually produce the same index, but migration-verifier will flag them as mismatches. Eventually we hope to fix that, but for now false positives are safer than incomplete checks.

This does teach index verification to ignore type differences, which is a common source of false-positives in verifications. For example, `{_id: Double(1)}` vs. `{_id: Long(1)}` will no longer be flagged as a mismatch. (NB: Due to this change, the metadata DB must now be on a v6+ cluster.)

This also replaces use of mongo.CollectionSpecification with a struct that allows detection of unrecognized fields. Thus we’ll avoid a similar problem for collection specifications. 

This also copies over mongosync’s Option implementation.