Add a worker tracker to migration-verifier’s progress logs. (#60)
This changeset adds a table of worker thread statuses to the progress logs. This will allow at-a-glance monitoring of how long different tasks have been in progress and will usefully surface any hung threads.

This also removes the “No tasks found. Sleeping.” log messages, which convey little but add a lot of clutter.