REP-5550 Add resumability for recheck generations (#85)
Previously migration-verifier did not persist its generation, so if it crashed in a generation after generation 0 it would still restart at generation 0.

This could cause any enqueued rechecks not to happen until the restarted verifier reached the prior verifier run’s generation. If the user called `writesOff` before then, though, the rechecks would never happen. Thus, migration-verifier could neglect important rechecks.

This fixes that by persisting migration-verifier’s generation at the start of each generation. Thus, any tasks that need to be rerun, or enqueued rechecks that need to be processed into checks for the next generation, will be replayed in the event of a crash.