Ignore/log `modify` events on the destination. (#97)
When mongosync migrates capped collections it sets the collection limits during commit, which causes the destination to emit `modify` events in its change stream. These have caused migration-verifier to fail because of the “unrecognized” event type.

This changeset alters that logic so that migration-verifier will log `modify` events from the destination and ignore them rather than treating them as fatal.

Additionally, this updates the unknown-event error to include the change event in its bson.Raw form rather than a struct. This ensures that we show the entire event rather than just those parts that we parse.