REP-5904 Switch to unordered inserts for persisting rechecks.
This appears to be significantly faster than upserting. It entails addition
of nontrivial logic to check for (simple) duplicate-key errors.

Since it means that we no longer update existing rechecks’ persisted
data sizes, this also updates tests not to check that part of the
recheck document. It’s fairly unimportant anyway since all it’s there
for is to try to optimize concurrency.