REP-5904 Refine computation of BSON length in recheck task generation
Previously we approximated the length of document IDs in recheck task
documents based on the ID values. This wasn’t good enough, though, for
cases with lots of small IDs.

This changeset refines that by having the relevant logic determine the
actual byte length of the encoded array of IDs.

The test for this change seems to cause GitHub Actions to terminate
testing. To avoid that, this test is skipped in CI.