REP-5996 Allow configurable max recheck task size; reduce default. (#109)
This allows users to set a custom maximum size for recheck tasks. The verifier observes this setting in 2 ways:
1. When creating recheck tasks, it will move onto the next task if the list of IDs has reached the maximum size.
2. When running recheck tasks, it splits the list of IDs into groups and runs each group separately (i.e., in series). This is so that, if a verification finishes generation 0 then hits problems in generation 1, the user can adjust the recheck size and have it take effect without restarting verification.

The reason for this change is that, as seen in HELP-73941, if the verifier sends many large queries to the server at once, it can deplete the server’s available memory, which can cause OOMs and elections. By throttling this new setting, customers can (hopefully) mitigate that problem.

This also reduces the default recheck size from 12 MiB to 8 MiB.