REP-6088 Tolerate high numbers of mismatches (#117)
Previously all document mismatches were recorded directly in the verification task. This meant, though, that if a task encompassed a large number of mismatched or missing documents, the verifier could fail to persist all of the mismatches, which caused a crash.

(The usual cause of excess mismatched/missing documents is starting migration-verifier before initial sync finishes, but it can also reasonably happen without REP-6129’s fix for queries against pre-v5 servers. See HELP-75910.)

This changeset makes the verifier save mismatches to a dedicated collection instead, one document per mismatch.

This change upends some familiar workflows for investigating mismatches: it’s no longer sufficient just to query the `verification_tasks` collection for mismatch information since the actual mismatches are recorded in a separate collection. To address this, the documentation now gives an aggregation pipeline that yields a similarly-useful result.

This entails a metadata version change. Because that’s happening, this also changes the task type `verify` to `verifyDocuments`. (That required some sorting workarounds in tests, which were tight-coupled to the task type strings.)