REP-6380 Add optional hash-based verification. (#122)
This replaces the current `--ignoreFieldOrder` flag with a new `--docCompareMethod` parameter that accepts 3 values: the 2 old modes, and a new one that compares documents via $toHashedIndexKey.

This is not set as the default because of the precision issues described in the README’s new section. The performance gains, though, probably outweigh those concerns for most migrations.

In a reference test, an initial scan that took 37 minutes with `binary` comparison took under 15 minutes with the new method.