Allow arbitrary DDL on the destination. (#125)
Some DDL operations are internal to a migration. We currently allow `modify` events for so that capped collections can have their limits set at the end of a migration. We also need to allow `createIndexes` to accommodate migrations where the destination’s indexes are created after initial sync.

Rather than allowing specific events, though, this changeset generalizes tolerance of all unknown events—which should be DDL—on the destination. Any time such an event appears in the destination’s change stream, we now log about it then discard it. We still fail if a DDL event happens on the source. This allows migrator tools flexibility in when they create collections or indexes while still making verification fail if an unsupported event happens on the source.