REP-6503 Avoid $type in partition queries unless needed (#129)
The $type checks that REP-6129 (PR #118) added to fix type bracketing have impeded performance against 4.4 clusters.

This changeset makes partition queries use $type only when the range requires it, which should only be true for a handful of partitions anyhow. The queries are simplified, too, so that instead of ANDing together predicates like `_id >= 12 OR $type in ["string", "symbol", ...])` and `id <= 99 OR $type in ["null", "minkey"]` we just have a single $type invocation (if it’s even there, which it usually won’t be now).

The existing checks for type-bracketing accommodation are expanded a bit here as well.