Avoid “decreasing” source docs count in recheck generations.
The aggregation to compile namespace stats assumes that
`source_documents_count` always represents the number of documents to
expect on the source. For recheck tasks, though, once finished the verifier
sets that field to the number of documents actually _found_ on the source.
Thus, as the verifier finishes tasks in recheck generations, the total
number of source documents shown in the progress reports decreases. This
is confusing.

This changeset rectifies that by always using the size of the `_ids` array
to count the # of source documents to report in recheck generations’s
progress reports.