REP-6634 Add time-series support (#140)
This changeset adds basic time-series support.

Because of the verifier’s `_id` dependency, this verifies buckets rather than logical measurements. This implies a requirement that the migration copy time-series collections via buckets as well since a logical replication would not preserve artifacts like bucket `_id`, and possibly not even the grouping of measurements.

Because of the bucket-level verification, details in mismatch reports are not very useful for time-series because they reference bucket-level fields that the logical API doesn’t expose.

This works with per-shard verification (i.e., it can verify with or without a view). It _does not_ currently support namespace filtering.