Track change stream stats independently of generation (#145)
Currently we compute change stream speed by looking at the number of events received in the current generation. This can produce inaccurate results when generations are short.

This changeset switches that rate computation to be tracked independently of the generation.

Logging and documentation are also revised to be more helpful regarding change stream lag.