Tolerate missing ns.coll in change event. (#149)
Missing `ns.coll` can happen if, e.g., there’s a `dropDatabase` event. This changeset tolerates that. (We don’t support such events on the source, but on the destination we ignore them.)

The fix rewrites the UnmarshalBSON method to avoid LookupErr. Pprof shows it slightly more efficient to do things this way, and it better mimics the driver’s own tolerance missing fields when unmarshaling.