REP-6735 Allow rechecks to be enqueued during recheck task gen (#157)
Historically, the verifier blocked the enqueuing of rechecks while generating new recheck tasks. This slowed down the change stream because it couldn’t enqueue rechecks during that time.

There seems to be no reason why the change stream needs to be blocked while creating new tasks, though. This changeset removes the restriction. It also tweaks the generation-numbering so that the generation persisted with the recheck is the generation where the document will be rechecked rather than the generation where the recheck was enqueued.

The net effect here is that, with the change stream not blocked for several seconds between generations, higher throughput is possible.