Optimize progress report queries (#167)
Previously, when running queries for the progress reports, verifier fetched all mismatches for all failed tasks. It would only display up to 40 mismatches, though … so even if there were thousands of mismatches, we discarded all but 40.

This changeset alters the logic so that 3 queries run in parallel:
- fetch first 20 differing-content mismatches
- fetch first 20 missing/changed mismatches
- count all mismatches, dividing them into differing-content vs. missing/changed

Thus, the server only returns & parses up to 40 documents for these queries, which is much lighter than returning every single mismatch for all tasks.