REP-6804 Refactor change stream reader to generic “change reader” (#171)
This refactors the change stream reader to be an implementation of a generic “change reader”. Some renames are made here (e.g., “handler” to “persistor”), which hopefully will lend greater clarity long-term.

This anticipates creation of an “oplog reader” that will replace the change stream reader in some use cases.