REP-6804 Simplify change handling logic (#173)
Previously the code maintained:
- an Eventual for the error in persisting rechecks
- an Eventual for the error in reading changes
- an empty channel to indicate when the change reader is done

This is overly complex.

This changeset simplifies the above. Now there is a single Eventual that holds the final status of both readers & both persistors. Together the reader & persistor threads are called “change handling”.

This also fixes a small bug in Eventual: previously it was impossible to store a nil value because Eventual was implemented via Option. Now it’s possible to store nil.