REP-6832 Always read latest-possible version of documents (#177)
This improves the prevention of stale reads to incorporate the latest-seen cluster time and to use majority read concern.

This is particularly relevant when reading from secondaries, but it can also, because of [SERVER-53813](https://jira.mongodb.org/browse/SERVER-53813), affect reads from primaries.

This removes a `lag` from a trace-level log in order to simplify things a bit. That data point is redundant with the periodically-reported lag in the logs anyway.