REP-6873 Handle nonempty initial change stream (#183)
Verifier has historically assumed that a change stream’s initial response will be empty (save for the resume token), but that’s not always the case.

When a nonempty initial response arrives, Verifier has been crashing because the driver’s ResumeToken() method has returned nil.

This changeset fixes Verifier to anticipate a nonempty initial response.

A generic cursor-batch-reader is added to mmongo to facilitate this. Also, the log about the start timestamp now more clearly distinguishes when the cluster time predates the resume token’s timestamp.