REP-6875 Make ChangeReader distinguish last-handled & last operation times. (#194)
Previously the ChangeReader exposed only the reader lag, but external tooling needs to track not merely the lag but also the reader’s exact last-handled timestamp.

This changeset reworks the ChangeReader accordingly so that we can add that information to the /progress endpoint in a later PR.

(This is an internal-only change that will facilitate further PRs for this JIRA ticket.