REP-6928 Make “last-handled” optime more meaningful (#200)
This changes (fixes?) the “last-handled” optime added in PR #199 to reflect the _rechecked_ optime.

The existing optime is updated when a recheck is enqueued. This indicates the change reader’s progress, but that doesn’t reflect the _verifier’s_ overall progress. What’s needed instead is a metric that’s updated after a recheck.

This changeset implements that.